Restrict `Matcher` class usage
